pyproject.toml
[tool.poetry]
name = "script merger"
version = "0.1.0"
description = "Merges text files into a word document"
authors = ["Matthew Pugh <pughmds@cyram.org>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
python-docx = "^0.8.11"
click = "^8.1.3"
python-decouple = "^3.8"


[tool.poetry.group.dev.dependencies]
black = "^23.9.1"
pylint = "^2.17.6"
pytest = "^7.4.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
